+++ src/bitcoind.cpp
@@ -3,9 +3,12 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
+#include "ui_interface.h"
+#include "init.h"
+#include "util.h"
+#include "main.h"
 #include "rpcserver.h"
 #include "rpcclient.h"
-#include "init.h"
 #include <boost/algorithm/string/predicate.hpp>
 
 void WaitForShutdown(boost::thread_group* threadGroup)
